requirements.txt
streamlit>=1.24.0
pdf2image>=1.16.3
Pillow>=10.0.0
openai>=1.0.0
python-docx
python-dotenv>=0.19.0
PyPDF2>=3.0.0